# packages/api-contract/openapi.yaml
openapi: 3.1.0
info:
  title: MateryalPH API
  version: 1.0.0-phase.1
  description: >-
    Versioned Phase 1 contract. FIN and MAT schemas below are foundations;
    later-phase operations are deliberately not exposed as placeholder routes.
servers:
  - url: /api/v1
...
paths:
...
  /auth/admin-invitations/accept:
    post:
      operationId: acceptAdminInvitation
      tags: [Authentication]
      security: [{ webCsrf: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AdminInvitationRequest" }
      responses:
        "201":
          description: Admin identity created; MFA enrollment is the next required step.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/RegistrationEnvelope" }
        "410": { $ref: "#/components/responses/ErrorResponse" }
components:
  securitySchemes:
    passportBearer: { type: http, scheme: bearer, bearerFormat: JWT }
    accessCookie: { type: apiKey, in: cookie, name: mp_access }
    mfaChallengeCookie: { type: apiKey, in: cookie, name: mp_mfa_challenge }
    botProofCookie: { type: apiKey, in: cookie, name: mp_bot_proof }
    webCsrf: { type: apiKey, in: header, name: X-CSRF-Token }
  responses:
    ErrorResponse:
      description: Safe structured error with X-Correlation-ID response header.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorEnvelope" }
...
  schemas:
    Meta: { type: object, additionalProperties: true }
    ApiError:
      type: object
      required: [code, message, details]
      properties:
        code: { type: string }
        message: { type: string }
        field: { type: [string, "null"] }
        correlation_id: { type: [string, "null"] }
        details: { type: object, additionalProperties: true }
    ErrorEnvelope:
      type: object
      required: [data, meta, errors]
      properties:
        data: { type: [object, "null"], additionalProperties: false }
        meta: { $ref: "#/components/schemas/Meta" }
        errors:
          type: array
          minItems: 1
          items: { $ref: "#/components/schemas/ApiError" }
    SuccessEnvelope:
      type: object
      required: [data, meta, errors]
      properties:
        data: { type: object, additionalProperties: true }
        meta: { $ref: "#/components/schemas/Meta" }
        errors: { type: array, maxItems: 0 }
...
    AdminInvitationRequest:
      type: object
      required: [token, full_name, password, password_confirmation, terms_accepted, privacy_accepted]
      properties:
        token: { type: string, writeOnly: true }
        full_name: { type: string, maxLength: 160 }
        password: { type: string, format: password, minLength: 14 }
        password_confirmation: { type: string, format: password }
        terms_accepted: { type: boolean, const: true }
        privacy_accepted: { type: boolean, const: true }
...
    RegistrationEnvelope:
      allOf:
        - $ref: "#/components/schemas/SuccessEnvelope"
        - type: object
          properties:
            data:
              type: object
              properties:
                user_id: { type: string, format: uuid }
                verification_required: { type: boolean }
                next: { type: string }
...